+++ test/qr_image_test.dart
@@ -6,11 +6,9 @@ void main() {
   late int moduleCount;
 
   setUpAll(() {
-    const typeNumber = 1;
     final qrCode = QrCode(
-      typeNumber,
-      QrErrorCorrectLevel.low,
-      QrPayload.fromData('test'),
+      payload: QrPayload.fromString('test'),
+      errorCorrectLevel: QrErrorCorrectLevel.low,
     );
     qrImage = QrImage(qrCode);
     moduleCount = qrImage.moduleCount;